float converting...

+++ ProjectWithGit/src/Project/BarLinesPDF.java
@@ -40,7 +40,7 @@ public class BarLinesPDF
 	
 	private static Phrase currentChar;
 	private static float noteFontSize = 6; //Size of the characters to be written to the page
-	private static float givenSpacing = (float)DataToArray.getSpacing(); //The spacing given at the start of the program, change to variable once we read it in
+	private static float givenSpacing = DataToArray.getSpacing(); //The spacing given at the start of the program, change to variable once we read it in
 	private static float barSpacing = 7; //Space between individual lines to be drawn
 	private static float whiteSpace = 1; //Space around a written number that does not have a bar line
 	private static float groupBarSpacing = 75; //Spaces between the groups of 6 lines
